Remove interface that's only used in one class
