+++ src/main/java/org/javarosa/core/model/TriggerableDag.java
@@ -100,7 +100,7 @@ public class TriggerableDag {
 
     private boolean predicateCaching = true;
     private final PredicateFilter cachingPredicateFilter = new CompareChildToAbsoluteExpressionFilter();
-    private final PredicateFilter indexPredicateFilter = new IndexPredicateFilter(new InMemTreeReferenceIndex());
+    private final PredicateFilter indexPredicateFilter = new IndexPredicateFilter();
 
     TriggerableDag(EventNotifierAccessor accessor) {
         this.accessor = accessor;
